+++ README.md
@@ -1,58 +1,36 @@
-# CodeWaveAssetCompetition2024
-网易低代码开发者活动-资产大赛2024
-![image](https://github.com/netease-lcap/CodeWaveAssetCompetition2024/assets/158463965/ea58284c-5be6-4b2d-956b-7b300499ee1e)
+# 依赖库名称
 
-欢迎开发者们参加2024网易低代码资产共建大赛，活动官网介绍：https://sf.163.com/2024lcapassetcompetition
+依赖库提供xml转json，在json获得特定Key的内容。
 
-参赛指引：
-![image](https://github.com/netease-lcap/CodeWaveAssetCompetition2024/assets/158463965/8f658d1c-0f81-4062-b794-a3d0e0ec900a)
+## 逻辑详情
 
+### convertXmlToJson
+将xml内容转json
+1. 入参：String：xml内容
+2. 出参：String: json格式,当返回错误，返回null并且打印error日志
 
-—大赛介绍
+### getXPathKey
+通过JSONPath 表达式在json内容获得特定的KEY 
+实例： 
+json: "{"name":"John","age":30,"address":{"city":"Hangzhou"}}"
+key: $.name -> 表示顶层的 name key
+key: $.address.city -> 表示嵌套在 address 中的 city key
+1. 入参：String：json内容
+2. 入参2: String: key JSONPath表达式
+3. 出参：String: 返回对应key的内容,当json或者key为空，返回null，当key格式错误返回null,并且打印错误日志
 
-1、资产共建大赛简介：
+## 使用步骤说明
 
-基于平台现有扩展集成能力，和开发者一起构建CodeWave平台资产市场，包括连接器、依赖库，打造丰富的开发场景资产，降低开发成本提升效率。活动结束后资产将会在平台的【资产中心】上架开放，提供企业&开发者后续使用。
+1. 应用引用依赖库
+![img_1.png](img_1.png)
+2. 逻辑调用示例截图
 
-2、资产共建大赛特色
+![img.png](img.png)
+![img_4.png](img_4.png)
 
-  1）赛道自由选择，大赛提供了三个赛道方向可供参赛开发者选择，包括命题组的连接器需求、依赖库需求和非命题组（开发者自由提供优质资产)；
-
-  2）赛事赋能保障，配套赋能直播课程和参赛交流社群，保障大家的学习、交流和问题咨询；
-
-  3）多劳多得，上不封顶，活动鼓励开发者可多次参与提供多个资产，验收通过后均可获得激励；
-
-  4）礼品随心选，活动以严选卡的形式发放激励，开发者可以根据自己的爱好前往严选平台购买心怡礼品
-
-
-—赛事规则
-
-1、参赛机制：
-
-  1）一个命题可允许最多5名开发者参加，开发者需要在命题需求发布帖子下方评论区进行报名领取需求；
-
-  2）一个命题涉及到多个资产提交，根据资产的质量度进行梯队评估激励，可多人同时获得一个需求的激励；
-
-  3）一个开发者最多同时参加三个命题需求领取，需求开发完提交验收通过后，会释放领取额度继续申领新需求；
-
-  4）一个开发者在活动期间超过三次占用命题参加额度，但未提交资产，则后续不可以参与资产命题需求的申领（自由资产提供不限制）；
-
-2、验收机制
-![image](https://github.com/netease-lcap/CodeWaveAssetCompetition2024/assets/158463965/1b02393f-3df9-41f5-a187-64e71f7cef9f)
-
-3、赛事激励：
-
-激励内容：严选卡
-
-说明：本次活动的激励因为资产价值恒量的不确定性，不再通过具体实物进行奖励，而是发放等同价值的严选卡。开发者获得严选卡激励后，可以前往严选平台绑定卡片自行进行商品挑选。 
-
-激励形式：
-
-  1）平台发布资产需求会提供对应的单个命题需求的激励额度池，开发者交付资产验收后符合规则即可获得一定的激励额度；
-
-  2）开发者自己向平台主动提供的资产，平台会验收评审资产的价值，给到对应的激励额度；
-
-严选平台优质商品推荐：https://docs.popo.netease.com/lingxi/65c7da790e5d40baaf7ee15e62e843e5
+![img_2.png](img_2.png)
+![img_3.png](img_3.png)
 
 
-
+## 应用演示链接
+https://dev-testlib1-nt.app.codewave.163.com/test_lib2
