+++ jar/nasl-metadata-maven-plugin-1.3.0/install.bat
@@ -1,0 +1,1 @@
+mvn install:install-file -Dfile="nasl-metadata-maven-plugin-1.3.0.jar" -DpomFile="pom.xml"
