--- 前端readme模板
@@ -1,40 +1,0 @@
-# 依赖库名称
-
-**依赖库设计**
-
-这个依赖库旨在提供一套用于低代码开发的工具集，以简化开发人员的工作流程。它包含了多个组件和逻辑，每个组件和逻辑都提供了特定的功能，使开发过程更加高效。
-
-**主要特性**
-
-- **组件一/逻辑一：** 提供功能一，功能二，功能三等多种功能点，适用于各种场景。
-- **组件二/逻辑二：** 具备功能四，功能五，功能六等多种特性，可满足不同的需求。
-
-## 组件一/逻辑一
-依次描述每个组件（逻辑）的特性
-
-**特性 1：** 描述特性 1 的功能和用途。
-
-**特性 2：** 描述特性 2 的功能和用途。
-
-...
-
-## 使用说明
-
-### 组件（组件名）
-
-- **attrs**
-  - attrName: 描述属性的作用和数据类型
-
-- **methods**
-  - methodName(params): 描述方法的作用和用法和返回值类型
-
-- **events**
-  - eventName: 描述事件的触发时机和用途
-
-### 逻辑（逻辑名）
-
-- **methodName(params):** 描述方法的作用和用法和返回值类型，传入参数的类型和参数是否为必填
-
-## 应用演示链接
-
-[查看示例演示](示例演示链接)
