--- 后端readme模板
@@ -1,25 +1,0 @@
-# 依赖库名称
-
-简短地介绍本依赖库做了什么，包含了哪些逻辑。可以突出描述一下创新点。
-
-## 逻辑详情
-
-### 逻辑一
-
-描述逻辑的功能，逻辑的出入参。
-
-### 逻辑二
-
-### ...
-
-## 使用步骤说明
-
-1.  应用引用依赖库
-2.  配置应用配置参数（如果有的话）
-3.  逻辑调用示例截图
-4.  ..
-
-## 应用演示链接
-
-使用了本依赖库的应用的链接。
-
