--- 资产共建大赛提交内容说明.md
@@ -1,8 +1,0 @@
-# 资产共建大赛提交内容说明
-
-1. 本次提交以PR方式提交
-2. 需要包含项目原始工程(源码)
-3. 源码工程中，除源码外，还要有README.md(说明)
-4. README.md参考如下
-   * [前端README](https://github.com/netease-lcap/CodeWaveAssetCompetition2024/blob/main/%E5%89%8D%E7%AB%AFreadme%E6%A8%A1%E6%9D%BF)
-   * [后端README](https://github.com/netease-lcap/CodeWaveAssetCompetition2024/blob/main/%E5%90%8E%E7%AB%AFreadme%E6%A8%A1%E6%9D%BF)
